fix: join create and update task handlers
