+++ Doppler.Sap/Enums/SapTaskEnum.cs
@@ -2,9 +2,8 @@ namespace Doppler.Sap.Enums
 {
     public enum SapTaskEnum
     {
-        CreateBusinessPartner = 1,
         BillingRequest = 2,
         CurrencyRate = 3,
-        UpdateBusinessPartner = 4
+        CreateOrUpdateBusinessPartner = 4
     }
 }
