+++ Doppler.Sap/Factory/CreateOrUpdateBusinessPartnerHandler.cs
@@ -11,13 +11,13 @@ using System.Threading.Tasks;
 
 namespace Doppler.Sap.Factory
 {
-    public class UpdateBusinessPartnerHandler
+    public class CreateOrUpdateBusinessPartnerHandler
     {
         private readonly SapConfig _sapConfig;
         private readonly ISapTaskHandler _sapTaskHandler;
         private readonly IHttpClientFactory _httpClientFactory;
 
-        public UpdateBusinessPartnerHandler(
+        public CreateOrUpdateBusinessPartnerHandler(
             IOptions<SapConfig> sapConfig,
             ISapTaskHandler sapTaskHandler,
             IHttpClientFactory httpClientFactory)
@@ -31,6 +31,46 @@ namespace Doppler.Sap.Factory
         {
             dequeuedTask = await _sapTaskHandler.CreateBusinessPartnerFromDopplerUser(dequeuedTask);
 
+            if (string.IsNullOrEmpty(dequeuedTask.ExistentBusinessPartner.FederalTaxID))
+            {
+                return await CreateBusinessPartner(dequeuedTask);
+            }
+            else
+            {
+                return await UpdateBusinessPartner(dequeuedTask);
+            }
+        }
+
+        private async Task<SapTaskResult> CreateBusinessPartner(SapTask dequeuedTask)
+        {
+            var message = new HttpRequestMessage()
+            {
+                RequestUri = new Uri($"{_sapConfig.BaseServerUrl}BusinessPartners/"),
+                Content = new StringContent(JsonConvert.SerializeObject(dequeuedTask.BusinessPartner),
+                    Encoding.UTF8,
+                    "application/json"),
+                Method = HttpMethod.Post
+            };
+
+            var cookies = await _sapTaskHandler.StartSession();
+            message.Headers.Add("Cookie", cookies.B1Session);
+            message.Headers.Add("Cookie", cookies.RouteId);
+
+            var client = _httpClientFactory.CreateClient();
+            var sapResponse = await client.SendAsync(message);
+
+            var taskResult = new SapTaskResult
+            {
+                IsSuccessful = sapResponse.IsSuccessStatusCode,
+                SapResponseContent = await sapResponse.Content.ReadAsStringAsync(),
+                TaskName = "Creating Business Partner"
+            };
+
+            return taskResult;
+        }
+
+        private async Task<SapTaskResult> UpdateBusinessPartner(SapTask dequeuedTask)
+        {
             //SAP uses a non conventional patch where you have to send only the fields that you want to be changed with the new values
             dequeuedTask.BusinessPartner.BPAddresses = GetBPAddressesPatchObject(dequeuedTask.BusinessPartner.BPAddresses);
             dequeuedTask.BusinessPartner.ContactEmployees = GetContactEmployeesPatchObject(dequeuedTask.BusinessPartner.ContactEmployees, dequeuedTask.ExistentBusinessPartner.ContactEmployees);
@@ -55,12 +95,13 @@ namespace Doppler.Sap.Factory
             var message = new HttpRequestMessage()
             {
                 RequestUri = new Uri($"{_sapConfig.BaseServerUrl}BusinessPartners('{dequeuedTask.ExistentBusinessPartner.CardCode}')"),
-                Content = new StringContent(JsonConvert.SerializeObject(dequeuedTask.BusinessPartner, new JsonSerializerSettings
-                {
-                    NullValueHandling = NullValueHandling.Ignore
-                })
-                        , Encoding.UTF8
-                        , "application/json"),
+                Content = new StringContent(JsonConvert.SerializeObject(dequeuedTask.BusinessPartner,
+                    new JsonSerializerSettings
+                    {
+                        NullValueHandling = NullValueHandling.Ignore
+                    }),
+                    Encoding.UTF8,
+                    "application/json"),
                 Method = HttpMethod.Patch
             };
