+++ Doppler.Sap/Factory/SapTaskFactory.cs
@@ -22,10 +22,8 @@ namespace Doppler.Sap.Factory
                     return await ((SetCurrencyRateHandler)_serviceProvider.GetService(typeof(SetCurrencyRateHandler))).Handle(sapTask);
                 case SapTaskEnum.BillingRequest:
                     return await ((BillingRequestHandler)_serviceProvider.GetService(typeof(BillingRequestHandler))).Handle(sapTask);
-                case SapTaskEnum.CreateBusinessPartner:
-                    return await ((CreateBusinessPartnerHandler)_serviceProvider.GetService(typeof(CreateBusinessPartnerHandler))).Handle(sapTask);
-                case SapTaskEnum.UpdateBusinessPartner:
-                    return await ((UpdateBusinessPartnerHandler)_serviceProvider.GetService(typeof(UpdateBusinessPartnerHandler))).Handle(sapTask);
+                case SapTaskEnum.CreateOrUpdateBusinessPartner:
+                    return await ((CreateOrUpdateBusinessPartnerHandler)_serviceProvider.GetService(typeof(CreateOrUpdateBusinessPartnerHandler))).Handle(sapTask);
                 default:
                     throw new ArgumentOutOfRangeException();
             }
