+++ Doppler.Sap/Startup.cs
@@ -83,8 +83,7 @@ namespace Doppler.Sap
             services.Configure<SapConfig>(Configuration.GetSection(nameof(SapConfig)));
             services.AddTransient<SetCurrencyRateHandler>();
             services.AddTransient<BillingRequestHandler>();
-            services.AddTransient<CreateBusinessPartnerHandler>();
-            services.AddTransient<UpdateBusinessPartnerHandler>();
+            services.AddTransient<CreateOrUpdateBusinessPartnerHandler>();
             services.AddTransient<ISapTaskFactory, SapTaskFactory>();
             services.AddSingleton<IDateTimeProvider, DateTimeProvider>();
             services.AddTransient<ISlackService, SlackService>();
